+++ BE_tasteal/Entity/DTO/Response/RecipeRes.cs
@@ -6,7 +6,7 @@ namespace BE_tasteal.Entity.DTO.Response
     {
         public int id { get; set; }
         public string? name { get; set; }
-        public float? rating { get; set; }
+        public decimal? rating { get; set; }
         public int? totalTime { get; set; }
         public int serving_size { get; set; }
         public string? introduction { get; set; }
