+++ BE_tasteal/Entity/Entity/RecipeEntity.cs
@@ -13,7 +13,7 @@ namespace BE_tasteal.Entity.Entity
         [Required]
         [MaxLength(255)]
         public string? name { get; set; }
-        public float? rating { get; set; }
+        public decimal? rating { get; set; }
         public int? totalTime { get; set; }
         public DateTime? active_time { get; set; }
         public int serving_size { get; set; }
